rename: DiscoPoP.h --> DiscoPoP.hpp
